Resolve pubkey/keypair args in a separate module
