+++ stake-accounts/src/main.rs
@@ -1,16 +1,15 @@
+mod arg_parser;
 mod args;
 mod stake_accounts;
 
+use crate::arg_parser::parse_args;
 use crate::args::{
-    parse_args, AuthorizeCommandConfig, Command, MoveCommandConfig, NewCommandConfig,
+    resolve_command, AuthorizeCommandConfig, Command, MoveCommandConfig, NewCommandConfig,
     RebaseCommandConfig,
 };
-use clap::ArgMatches;
-use solana_clap_utils::keypair::{pubkey_from_path, signer_from_path};
 use solana_cli_config::Config;
 use solana_client::client_error::ClientError;
 use solana_client::rpc_client::RpcClient;
-use solana_remote_wallet::remote_wallet::{maybe_wallet_manager, RemoteWalletManager};
 use solana_sdk::{
     message::Message,
     native_token::lamports_to_sol,
@@ -21,63 +20,6 @@ use solana_sdk::{
 };
 use std::env;
 use std::error::Error;
-use std::sync::Arc;
-
-fn resolve_stake_authority(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Box<dyn Signer>, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    signer_from_path(&matches, key_url, "stake authority", wallet_manager)
-}
-
-fn resolve_withdraw_authority(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Box<dyn Signer>, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    signer_from_path(&matches, key_url, "withdraw authority", wallet_manager)
-}
-
-fn resolve_new_stake_authority(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Pubkey, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    pubkey_from_path(&matches, key_url, "new stake authority", wallet_manager)
-}
-
-fn resolve_new_withdraw_authority(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Pubkey, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    pubkey_from_path(&matches, key_url, "new withdraw authority", wallet_manager)
-}
-
-fn resolve_fee_payer(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Box<dyn Signer>, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    signer_from_path(&matches, key_url, "fee-payer", wallet_manager)
-}
-
-fn resolve_base_pubkey(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Pubkey, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    pubkey_from_path(&matches, key_url, "base pubkey", wallet_manager)
-}
-
-fn resolve_new_base_keypair(
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    key_url: &str,
-) -> Result<Box<dyn Signer>, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    signer_from_path(&matches, key_url, "new base pubkey", wallet_manager)
-}
 
 fn get_balance_at(client: &RpcClient, pubkey: &Pubkey, i: usize) -> Result<u64, ClientError> {
     let address = stake_accounts::derive_stake_account_address(pubkey, i);
@@ -105,77 +47,43 @@ fn get_balances(
 
 fn process_new_stake_account(
     client: &RpcClient,
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    new_config: &NewCommandConfig,
-) -> Result<Signature, Box<dyn Error>> {
-    let matches = ArgMatches::default();
-    let fee_payer_keypair = resolve_fee_payer(wallet_manager, &new_config.fee_payer)?;
-    let funding_keypair = signer_from_path(
-        &matches,
-        &new_config.funding_keypair,
-        "funding keypair",
-        wallet_manager,
-    )?;
-    let base_keypair = signer_from_path(
-        &matches,
-        &new_config.base_keypair,
-        "base keypair",
-        wallet_manager,
-    )?;
-    let stake_authority_pubkey = pubkey_from_path(
-        &matches,
-        &new_config.stake_authority,
-        "stake authority",
-        wallet_manager,
-    )?;
-    let withdraw_authority_pubkey = pubkey_from_path(
-        &matches,
-        &new_config.withdraw_authority,
-        "withdraw authority",
-        wallet_manager,
-    )?;
+    config: &NewCommandConfig<Pubkey, Box<dyn Signer>>,
+) -> Result<Signature, ClientError> {
     let message = stake_accounts::new_stake_account(
-        &fee_payer_keypair.pubkey(),
-        &funding_keypair.pubkey(),
-        &base_keypair.pubkey(),
-        new_config.lamports,
-        &stake_authority_pubkey,
-        &withdraw_authority_pubkey,
-        new_config.index,
+        &config.fee_payer.pubkey(),
+        &config.funding_keypair.pubkey(),
+        &config.base_keypair.pubkey(),
+        config.lamports,
+        &config.stake_authority,
+        &config.withdraw_authority,
+        config.index,
     );
-    let signers = vec![&*fee_payer_keypair, &*funding_keypair, &*base_keypair];
+    let signers = vec![
+        &*config.fee_payer,
+        &*config.funding_keypair,
+        &*config.base_keypair,
+    ];
     let signature = send_message(client, message, &signers)?;
     Ok(signature)
 }
 
 fn process_authorize_stake_accounts(
     client: &RpcClient,
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    authorize_config: &AuthorizeCommandConfig,
-) -> Result<(), Box<dyn Error>> {
-    let fee_payer_keypair = resolve_fee_payer(wallet_manager, &authorize_config.fee_payer)?;
-    let base_pubkey = resolve_base_pubkey(wallet_manager, &authorize_config.base_pubkey)?;
-    let stake_authority_keypair =
-        resolve_stake_authority(wallet_manager, &authorize_config.stake_authority)?;
-    let withdraw_authority_keypair =
-        resolve_withdraw_authority(wallet_manager, &authorize_config.withdraw_authority)?;
-    let new_stake_authority_pubkey =
-        resolve_new_stake_authority(wallet_manager, &authorize_config.new_stake_authority)?;
-    let new_withdraw_authority_pubkey =
-        resolve_new_withdraw_authority(wallet_manager, &authorize_config.new_withdraw_authority)?;
+    config: &AuthorizeCommandConfig<Pubkey, Box<dyn Signer>>,
+) -> Result<(), ClientError> {
     let messages = stake_accounts::authorize_stake_accounts(
-        &fee_payer_keypair.pubkey(),
-        &base_pubkey,
-        &stake_authority_keypair.pubkey(),
-        &withdraw_authority_keypair.pubkey(),
-        &new_stake_authority_pubkey,
-        &new_withdraw_authority_pubkey,
-        authorize_config.num_accounts,
+        &config.fee_payer.pubkey(),
+        &config.base_pubkey,
+        &config.stake_authority.pubkey(),
+        &config.withdraw_authority.pubkey(),
+        &config.new_stake_authority,
+        &config.new_withdraw_authority,
+        config.num_accounts,
     );
     let signers = vec![
-        &*fee_payer_keypair,
-        &*stake_authority_keypair,
-        &*withdraw_authority_keypair,
+        &*config.fee_payer,
+        &*config.stake_authority,
+        &*config.withdraw_authority,
     ];
     for message in messages {
         let signature = send_message(client, message, &signers)?;
@@ -186,29 +94,22 @@ fn process_authorize_stake_accounts(
 
 fn process_rebase_stake_accounts(
     client: &RpcClient,
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    rebase_config: &RebaseCommandConfig,
-) -> Result<(), Box<dyn Error>> {
-    let fee_payer_keypair = resolve_fee_payer(wallet_manager, &rebase_config.fee_payer)?;
-    let base_pubkey = resolve_base_pubkey(wallet_manager, &rebase_config.base_pubkey)?;
-    let new_base_keypair =
-        resolve_new_base_keypair(wallet_manager, &rebase_config.new_base_keypair)?;
-    let stake_authority_keypair =
-        resolve_stake_authority(wallet_manager, &rebase_config.stake_authority)?;
+    config: &RebaseCommandConfig<Pubkey, Box<dyn Signer>>,
+) -> Result<(), ClientError> {
     let addresses =
-        stake_accounts::derive_stake_account_addresses(&base_pubkey, rebase_config.num_accounts);
+        stake_accounts::derive_stake_account_addresses(&config.base_pubkey, config.num_accounts);
     let balances = get_balances(&client, addresses)?;
 
     let messages = stake_accounts::rebase_stake_accounts(
-        &fee_payer_keypair.pubkey(),
-        &new_base_keypair.pubkey(),
-        &stake_authority_keypair.pubkey(),
+        &config.fee_payer.pubkey(),
+        &config.new_base_keypair.pubkey(),
+        &config.stake_authority.pubkey(),
         &balances,
     );
     let signers = vec![
-        &*fee_payer_keypair,
-        &*new_base_keypair,
-        &*stake_authority_keypair,
+        &*config.fee_payer,
+        &*config.new_base_keypair,
+        &*config.stake_authority,
     ];
     for message in messages {
         let signature = send_message(client, message, &signers)?;
@@ -219,41 +120,28 @@ fn process_rebase_stake_accounts(
 
 fn process_move_stake_accounts(
     client: &RpcClient,
-    wallet_manager: Option<&Arc<RemoteWalletManager>>,
-    move_config: &MoveCommandConfig,
-) -> Result<(), Box<dyn Error>> {
+    move_config: &MoveCommandConfig<Pubkey, Box<dyn Signer>>,
+) -> Result<(), ClientError> {
     let authorize_config = &move_config.authorize_config;
-    let rebase_config = &move_config.rebase_config;
-    let fee_payer_keypair = resolve_fee_payer(wallet_manager, &authorize_config.fee_payer)?;
-    let base_pubkey = resolve_base_pubkey(wallet_manager, &authorize_config.base_pubkey)?;
-    let new_base_keypair =
-        resolve_new_base_keypair(wallet_manager, &rebase_config.new_base_keypair)?;
-    let stake_authority_keypair =
-        resolve_stake_authority(wallet_manager, &authorize_config.stake_authority)?;
-    let withdraw_authority_keypair =
-        resolve_withdraw_authority(wallet_manager, &authorize_config.withdraw_authority)?;
-    let new_stake_authority_pubkey =
-        resolve_new_stake_authority(wallet_manager, &authorize_config.new_stake_authority)?;
-    let new_withdraw_authority_pubkey =
-        resolve_new_withdraw_authority(wallet_manager, &authorize_config.new_withdraw_authority)?;
+    let config = &move_config.rebase_config;
     let addresses =
-        stake_accounts::derive_stake_account_addresses(&base_pubkey, authorize_config.num_accounts);
+        stake_accounts::derive_stake_account_addresses(&config.base_pubkey, config.num_accounts);
     let balances = get_balances(&client, addresses)?;
 
     let messages = stake_accounts::move_stake_accounts(
-        &fee_payer_keypair.pubkey(),
-        &new_base_keypair.pubkey(),
-        &stake_authority_keypair.pubkey(),
-        &withdraw_authority_keypair.pubkey(),
-        &new_stake_authority_pubkey,
-        &new_withdraw_authority_pubkey,
+        &config.fee_payer.pubkey(),
+        &config.new_base_keypair.pubkey(),
+        &config.stake_authority.pubkey(),
+        &authorize_config.withdraw_authority.pubkey(),
+        &authorize_config.new_stake_authority,
+        &authorize_config.new_withdraw_authority,
         &balances,
     );
     let signers = vec![
-        &*fee_payer_keypair,
-        &*new_base_keypair,
-        &*stake_authority_keypair,
-        &*withdraw_authority_keypair,
+        &*config.fee_payer,
+        &*config.new_base_keypair,
+        &*config.stake_authority,
+        &*authorize_config.withdraw_authority,
     ];
     for message in messages {
         let signature = send_message(client, message, &signers)?;
@@ -274,50 +162,45 @@ fn send_message<S: Signers>(
 
 fn main() -> Result<(), Box<dyn Error>> {
     let command_config = parse_args(env::args_os());
-    let config = Config::load(&command_config.config_file)?;
-    let json_rpc_url = command_config.url.unwrap_or(config.json_rpc_url);
+    let app_config = Config::load(&command_config.config_file)?;
+    let json_rpc_url = command_config.url.unwrap_or(app_config.json_rpc_url);
     let client = RpcClient::new(json_rpc_url);
 
-    let wallet_manager = maybe_wallet_manager()?;
-    let wallet_manager = wallet_manager.as_ref();
-    match command_config.command {
-        Command::New(new_config) => {
-            process_new_stake_account(&client, wallet_manager, &new_config)?;
+    match resolve_command(&command_config.command)? {
+        Command::New(config) => {
+            process_new_stake_account(&client, &config)?;
         }
-        Command::Count(count_config) => {
-            let base_pubkey = resolve_base_pubkey(wallet_manager, &count_config.base_pubkey)?;
-            let num_accounts = count_stake_accounts(&client, &base_pubkey)?;
+        Command::Count(config) => {
+            let num_accounts = count_stake_accounts(&client, &config.base_pubkey)?;
             println!("{}", num_accounts);
         }
-        Command::Addresses(query_config) => {
-            let base_pubkey = resolve_base_pubkey(wallet_manager, &query_config.base_pubkey)?;
+        Command::Addresses(config) => {
             let addresses = stake_accounts::derive_stake_account_addresses(
-                &base_pubkey,
-                query_config.num_accounts,
+                &config.base_pubkey,
+                config.num_accounts,
             );
             for address in addresses {
                 println!("{:?}", address);
             }
         }
-        Command::Balance(query_config) => {
-            let base_pubkey = resolve_base_pubkey(wallet_manager, &query_config.base_pubkey)?;
+        Command::Balance(config) => {
             let addresses = stake_accounts::derive_stake_account_addresses(
-                &base_pubkey,
-                query_config.num_accounts,
+                &config.base_pubkey,
+                config.num_accounts,
             );
             let balances = get_balances(&client, addresses)?;
             let lamports: u64 = balances.into_iter().map(|(_, bal)| bal).sum();
             let sol = lamports_to_sol(lamports);
             println!("{} SOL", sol);
         }
-        Command::Authorize(authorize_config) => {
-            process_authorize_stake_accounts(&client, wallet_manager, &authorize_config)?;
+        Command::Authorize(config) => {
+            process_authorize_stake_accounts(&client, &config)?;
         }
-        Command::Rebase(rebase_config) => {
-            process_rebase_stake_accounts(&client, wallet_manager, &rebase_config)?;
+        Command::Rebase(config) => {
+            process_rebase_stake_accounts(&client, &config)?;
         }
-        Command::Move(move_config) => {
-            process_move_stake_accounts(&client, wallet_manager, &move_config)?;
+        Command::Move(config) => {
+            process_move_stake_accounts(&client, &config)?;
         }
     }
     Ok(())
